+++ src/store/api/backend/index.ts
@@ -102,6 +102,8 @@ export {
   useRecommendPackageMutation,
 } from './hosted';
 export {
+  getHostArch,
+  getHostDistro,
   toComposerComposeRequest,
   useExportBlueprintCockpitQuery,
   useGetRegistryAuthStatusQuery,
