+++ src/store/api/backend/onprem/composerApi/helpers/index.ts
@@ -3,6 +3,7 @@ export { parseJsonUnsafe } from './parseJson';
 export { lookupDatastreamDistro } from './dataStreamLookup';
 export { getBlueprintsPath } from './getBlueprintsPath';
 export { getCloudConfigs } from './getCloudConfigs';
+export { getHostArch, getHostDistro } from './hostInfo';
 export {
   byDistroDescending,
   inferDistro,
@@ -13,4 +14,16 @@ export {
 } from './podman';
 export { readComposes } from './readComposes';
 export { safeReadJsonFile } from './safeReadJsonFile';
+export { mapHostedToOnPrem, mapOnPremToHosted } from './blueprintMapper';
+export type {
+  BlueprintOnPrem,
+  CustomizationsOnPrem,
+  CustomRepositoryOnPrem,
+  FileSystemOnPrem,
+  GroupOnPrem,
+  GroupsPackagesOnPrem,
+  PackagesOnPrem,
+  SshKeyOnPrem,
+  UserOnPrem,
+} from './blueprintMapper';
 export { toComposerComposeRequest } from './toComposerComposeRequest';
